+++ src/main/java/com/application/close/bridge/BridgeService.java
@@ -1,9 +1,0 @@
-package com.application.close.bridge;
-
-public interface BridgeService {
-
-	void connectModToMqtt(int modTcpId, int mqttParamId);
-
-	void disconnectModFromMqtt(int modTcpId);
-
-}