+++ src/main/java/com/application/close/helper/MemoryBuffer.java
@@ -18,14 +18,11 @@ public class MemoryBuffer {
 	// Maps a unique Mqtt param ID to its corresponding MqttClient instance.
 	private ConcurrentHashMap<Integer, MqttClient> mqttClient;
 
-	// Maps a unique Modbus data id to its corresponding Mqtt param ID.
-	private ConcurrentHashMap<Integer, Integer> modMqttBridge;
-
 	public MemoryBuffer() {
 		super();
 		this.modbusMaster = new ConcurrentHashMap<>();
 		this.mqttClient = new ConcurrentHashMap<>();
-		this.modMqttBridge = new ConcurrentHashMap<>();
+
 	}
 
 }