+++ src/main/java/com/application/close/modtcp/service/ModService.java
@@ -2,26 +2,26 @@ package com.application.close.modtcp.service;
 
 public interface ModService {
 
-	String connectToSlaveDevice(Integer tcpDataId);
+	String connectToSlaveDevice(int tcpDataId);
 
-	String reconnect(Integer tcpDataId);
+	String reconnect(int tcpDataId);
 
-	String disconnect(Integer tcpDataId);
+	String disconnect(int tcpDataId);
 
-	boolean[] readCoils(Integer tcpDataId, int slaveId, int offset, int quantity);
+	boolean[] readCoils(int tcpDataId, int slaveId, int offset, int quantity);
 
-	boolean[] readDiscreteInputs(Integer tcpDataId, int slaveId, int offset, int quantity);
+	boolean[] readDiscreteInputs(int tcpDataId, int slaveId, int offset, int quantity);
 
-	int[] readHoldingRegisters(Integer tcpDataId, int slaveId, int offset, int quantity);
+	int[] readHoldingRegisters(int tcpDataId, int slaveId, int offset, int quantity);
 
-	int[] readInputRegisters(Integer tcpDataId, int slaveId, int offset, int quantity);
+	int[] readInputRegisters(int tcpDataId, int slaveId, int offset, int quantity);
 
-	void writeSingleCoil(Integer tcpDataId, int slaveId, int offset, boolean value);
+	void writeSingleCoil(int tcpDataId, int slaveId, int offset, boolean value);
 
-	void writeMultipleCoils(Integer tcpDataId, int slaveId, int offset, boolean[] values);
+	void writeMultipleCoils(int tcpDataId, int slaveId, int offset, boolean[] values);
 
-	void writeSingleRegister(Integer tcpDataId, int slaveId, int offset, int value);
+	void writeSingleRegister(int tcpDataId, int slaveId, int offset, int value);
 
-	void writeMultipleRegisters(Integer tcpDataId, int slaveId, int offset, int[] values);
+	void writeMultipleRegisters(int tcpDataId, int slaveId, int offset, int[] values);
 
 }