+++ src/main/java/com/application/close/mqtt/service/MqttServiceImpl.java
@@ -9,11 +9,8 @@ import org.eclipse.paho.client.mqttv3.MqttMessage;
 import org.springframework.stereotype.Service;
 
 import com.application.close.exception.BadRequestException;
-import com.application.close.exception.ResourceNotFoundException;
 import com.application.close.helper.MemoryBuffer;
 import com.application.close.mqtt.entity.MqttParam;
-import com.application.close.mqtt.payload.MqttConnectResp;
-
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 
@@ -30,24 +27,16 @@ public class MqttServiceImpl implements MqttService {
 	private static final String WILL_MESSAGE = "Mqtt client: %s disconnected unexpectedly.";
 
 	@Override
-	public MqttConnectResp connect(int mqttParamId) {
+	public String connect(int mqttParamId) {
 		MqttParam param = paramService.getById(mqttParamId);
 
-		if (param == null) {
-			throw new ResourceNotFoundException("MqttParam", "mqttParamId", mqttParamId);
-		}
-
-		if (param.isConnected()) {
-			throw new BadRequestException("MQTT client is already connected for ID: " + mqttParamId);
+		if (buffer.getMqttClient().containsKey(mqttParamId)) {
+			return reconnect(mqttParamId);
 		}
 
 		if (param.isAuthEnabled()) {
 			throw new BadRequestException("Authentication must be disabled for a normal (non-auth) connection.");
 		}
-
-		MqttConnectResp resp = new MqttConnectResp();
-		resp.setClientId(param.getClientId());
-		resp.setMqttUrl(param.getUrl());
 
 		MqttClient mqttClient = null;
 		try {
@@ -69,7 +58,6 @@ public class MqttServiceImpl implements MqttService {
 				@Override
 				public void connectionLost(Throwable cause) {
 					log.warn("MQTT connection lost for clientId={}: {}", param.getClientId(), cause.getMessage());
-					paramService.updateConnectionStatus(mqttParamId, false);
 				}
 
 				@Override
@@ -79,32 +67,31 @@ public class MqttServiceImpl implements MqttService {
 					} else {
 						log.info("MQTT client connected for the first time to {}", serverURI);
 					}
-					paramService.updateConnectionStatus(mqttParamId, true);
-
 				}
 			});
 
 			mqttClient.connect(getMqttConnectOptions(param));
 			buffer.getMqttClient().put(mqttParamId, mqttClient);
 
-			resp.setConnected(true);
 		} catch (MqttException e) {
-			resp.setConnected(false);
-			log.error("MQTT connection failed for {}: {}", param.getUrl(), e.getMessage(), e);
+			try {
+				if (mqttClient != null)
+					mqttClient.close();
+			} catch (Exception ex) {
+				log.warn("Failed to close MQTT client after error: {}", ex.getMessage());
+			}
+			throw new BadRequestException(
+					String.format("MQTT connection failed for %s: %s", param.getUrl(), e.getMessage()));
 		}
-		return resp;
+		return String.format("Mqtt client connected for paramId: %s", mqttParamId);
 	}
 
 	@Override
-	public MqttConnectResp connectWithAuth(int mqttParamId) {
+	public String connectWithAuth(int mqttParamId) {
 		MqttParam param = paramService.getById(mqttParamId);
 
-		if (param == null) {
-			throw new ResourceNotFoundException("MqttParam", "mqttParamId", mqttParamId);
-		}
-
-		if (param.isConnected()) {
-			throw new BadRequestException("MQTT client is already connected for ID: " + mqttParamId);
+		if (buffer.getMqttClient().containsKey(mqttParamId)) {
+			return reconnect(mqttParamId);
 		}
 
 		if (!param.isAuthEnabled())
@@ -112,10 +99,6 @@ public class MqttServiceImpl implements MqttService {
 
 		if (param.getUserName() == null || param.getPassword() == null)
 			throw new BadRequestException("Username or password missing for Auth connection.");
-
-		MqttConnectResp resp = new MqttConnectResp();
-		resp.setClientId(param.getClientId());
-		resp.setMqttUrl(param.getUrl());
 
 		MqttClient mqttClient = null;
 		try {
@@ -135,7 +118,6 @@ public class MqttServiceImpl implements MqttService {
 				@Override
 				public void connectionLost(Throwable cause) {
 					log.warn("MQTT connection lost for clientId={}: {}", param.getClientId(), cause.getMessage());
-					paramService.updateConnectionStatus(mqttParamId, false);
 				}
 
 				@Override
@@ -145,7 +127,6 @@ public class MqttServiceImpl implements MqttService {
 					} else {
 						log.info("MQTT client connected for the first time to {}", serverURI);
 					}
-					paramService.updateConnectionStatus(mqttParamId, true);
 
 				}
 			});
@@ -157,12 +138,34 @@ public class MqttServiceImpl implements MqttService {
 			mqttClient.connect(mqttOptions);
 			buffer.getMqttClient().put(mqttParamId, mqttClient);
 
-			resp.setConnected(true);
 		} catch (MqttException e) {
-			resp.setConnected(false);
-			log.error("MQTT connection failed for {}: {}", param.getUrl(), e.getMessage(), e);
+
+			try {
+				if (mqttClient != null)
+					mqttClient.close();
+			} catch (Exception ex) {
+				log.warn("Failed to close MQTT client after error: {}", ex.getMessage());
+			}
+
+			throw new BadRequestException(
+					String.format("MQTT connection failed for %s: %s", param.getUrl(), e.getMessage()));
 		}
-		return resp;
+		return String.format("Mqtt client connected for paramId: %s", mqttParamId);
+	}
+
+	@Override
+	public String reconnect(int mqttParamId) {
+		MqttClient client = buffer.getMqttClient().get(mqttParamId);
+		try {
+			if (client.isConnected()) {
+				return "Mqtt client already active for mqttParamId: " + mqttParamId;
+			} else {
+				client.reconnect();
+				return "Reconnected Mqtt client for mqttParamId: " + mqttParamId;
+			}
+		} catch (MqttException e) {
+			throw new BadRequestException("Reconnect failed for mqttParamId " + mqttParamId + ": " + e.getMessage());
+		}
 	}
 
 	private MqttConnectOptions getMqttConnectOptions(MqttParam param) {
@@ -177,23 +180,21 @@ public class MqttServiceImpl implements MqttService {
 
 	@Override
 	public void disconnect(int mqttParamId) {
-		MqttParam param = paramService.getById(mqttParamId);
 
-		if (param == null)
-			throw new ResourceNotFoundException("MqttParam", "mqttParamId", mqttParamId);
-
-		if (!param.isConnected())
-			throw new BadRequestException("MQTT client is not connected");
+		MqttClient client = buffer.getMqttClient().get(mqttParamId);
+		if (client == null) {
+			throw new BadRequestException("No Mqtt Client found for mqttParamId: " + mqttParamId);
+		}
 
 		try {
-			if (buffer.getMqttClient().containsKey(mqttParamId)) {
-				buffer.getMqttClient().get(mqttParamId).disconnect();
-				buffer.getMqttClient().remove(mqttParamId);
-				paramService.updateConnectionStatus(mqttParamId, false);
-				log.info("MQTT client disconnected successfully for ID: {}", mqttParamId);
+			if (client.isConnected()) {
+				client.disconnect();
 			}
+
+			log.info("MQTT client disconnected successfully for paramId: {}", mqttParamId);
+
 		} catch (MqttException e) {
-			log.error("Failed to disconnect MQTT client with ID {}: {}", mqttParamId, e.getMessage());
+			log.error("Failed to disconnect MQTT client with paramId: {} , message: {}", mqttParamId, e.getMessage());
 		}
 	}